[touch] 调整触摸头文件至 rtdevice.h 中
